+++ common/Cargo.toml
@@ -12,7 +12,9 @@ edition = "2018"
 
 [dependencies]
 structopt = { version = "0.3.13", default_features = false }
-config = { version = "0.9.3" }
+config = { version = "0.9.3", default_features = false, features = ["toml"] }
+serde = { version = "1.0.106", default_features = false }
+serde_json = "1.0.51"
 dirs = "2.0"
 get_if_addrs = "0.5.3"
 log = "0.4.8"
@@ -24,3 +26,5 @@ sha2 = "0.8.0"
 [dev-dependencies]
 tempdir = "0.3.7"
 tari_test_utils = { version = "^0.0", path = "../infrastructure/test_utils"}
+serde = { version = "1.0.106", features = ["derive"] }
+anyhow = "1.0"
