+++ common/src/configuration/error.rs
